// src/client/app/_models/constante.ts
/**
 * Created by Felipe Aguirre - Jenniferth Escobar on 23/02/2017.
 */
export class Constante {
    idConstante: number;
    constante: string;
    idTipoDato: number;
    valor: string;
    descripcion: string;
    indicadorHabilitado: boolean = true;
    auditoriaUsuario: number;
    auditoriaFecha: Date;
}